Criação de 2 novos arquivos (home_controller e analyze_controller), que cuidarão das rotas especificas ao seu domínio.
